pyproject.toml
[tool.uv.workspace]
members = [
    "docketeer",
    "docketeer-1password",
    "docketeer-agentskills",
    "docketeer-anthropic",
    "docketeer-autonomy",
    "docketeer-bubblewrap",
    "docketeer-deepinfra",
    "docketeer-git",
    "docketeer-mcp",
    "docketeer-monty",
    "docketeer-rocketchat",
    "docketeer-search",
    "docketeer-subprocess",
    "docketeer-tui",
    "docketeer-web",
]

[dependency-groups]
dev = [
    "docketeer",
    "docketeer-1password",
    "docketeer-agentskills",
    "docketeer-anthropic",
    "docketeer-autonomy",
    "docketeer-bubblewrap",
    "docketeer-deepinfra",
    "docketeer-git",
    "docketeer-mcp",
    "docketeer-monty",
    "docketeer-rocketchat",
    "docketeer-search",
    "docketeer-subprocess",
    "docketeer-tui",
    "docketeer-web",
    "codespell",
    "loq",
    "prek",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-env",
    "pytest-timeout",
    "pytest-xdist",
    "respx>=0.22.0",
    "ruff",
    "ty",
]

[tool.uv.sources]
docketeer = { workspace = true }
docketeer-1password = { workspace = true }
docketeer-agentskills = { workspace = true }
docketeer-anthropic = { workspace = true }
docketeer-autonomy = { workspace = true }
docketeer-bubblewrap = { workspace = true }
docketeer-deepinfra = { workspace = true }
docketeer-git = { workspace = true }
docketeer-mcp = { workspace = true }
docketeer-monty = { workspace = true }
docketeer-rocketchat = { workspace = true }
docketeer-search = { workspace = true }
docketeer-subprocess = { workspace = true }
docketeer-tui = { workspace = true }
docketeer-web = { workspace = true }

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "F",        # Pyflakes
    "E",        # pycodestyle errors
    "W",        # pycodestyle warnings
    "I",        # isort
    "ANN001",   # missing-type-function-argument
    "ANN002",   # missing-type-args
    "ANN003",   # missing-type-kwargs
    "ANN201",   # missing-return-type-undocumented-public-function
    "ANN202",   # missing-return-type-private-function
    "ANN204",   # missing-return-type-special-method
    "ANN205",   # missing-return-type-staticmethod
    "ANN206",   # missing-return-type-classmethod
    "PGH003",   # blanket-type-ignore
    "TRY",      # tryceratops
    "S110",     # try-except-pass
    "S112",     # try-except-continue
    "PT",       # flake8-pytest-style
    "B",        # flake8-bugbear
    "SIM",      # flake8-simplify
    "UP",       # pyupgrade
    "PGH",      # pygrep-hooks
]

ignore = [
    "E501",     # line-too-long - using ruff format
    "TRY003",   # raise-vanilla-args
    "TRY002",   # raise-vanilla-class
    "TRY300",   # try-consider-else
    "TRY401",   # verbose-log-message
    "SIM117",   # multiple-with-statements
    "PT011",    # pytest-raises-too-broad
    "BLE001",   # blind-except
    "B904",     # raise-without-from-inside-except
    "B008",     # function-call-in-default-argument - Docket uses this
    "ANN401",   # any-type
]

[tool.ruff.lint.isort]
known-first-party = ["docketeer", "docketeer_1password", "docketeer_agentskills", "docketeer_anthropic", "docketeer_autonomy", "docketeer_bubblewrap", "docketeer_deepinfra", "docketeer_git", "docketeer_mcp", "docketeer_monty", "docketeer_rocketchat", "docketeer_search", "docketeer_subprocess", "docketeer_tui", "docketeer_web"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = ["ANN201"]

[tool.ruff.lint.flake8-annotations]
suppress-dummy-args = true
allow-star-arg-any = false

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = true
parametrize-names-type = "tuple"
parametrize-values-type = "list"
parametrize-values-row-type = "tuple"

[tool.coverage.run]
branch = true
omit = ["*/__main__.py", "*/testing.py"]

[tool.coverage.report]
precision = 2
skip_empty = true
skip_covered = true
exclude_also = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@overload",
    "@abstractmethod",
]

[tool.ty.rules]
unused-ignore-comment = "error"
invalid-ignore-comment = "error"
ignore-comment-unknown-rule = "error"
possibly-missing-attribute = "error"
possibly-missing-import = "error"
redundant-cast = "error"
